requirements.txt
# Core
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg[binary]

# Filtering & Search
django-filter==24.2

# CORS
django-cors-headers==4.3.1

# Environment config
django-environ==0.11.2

# Async task queue
celery==5.3.6
redis==5.0.4
django-celery-beat==2.6.0
django-celery-results==2.5.1

# API Docs
drf-spectacular==0.27.2

# Image handling
Pillow>=10.3.0

# Utilities
python-dateutil==2.9.0
django-extensions==3.2.3
